# user-service/src/main/java/com/user_service/user_service/dto/ChangePasswordRequestDTO.java
package com.user_service.user_service.dto;

import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
public class ChangePasswordRequestDTO {
    private String oldPassword;
    private String newPassword;
    private String confirmNewPassword;
}